# openapi.yaml
openapi: 3.0.1
info:
  title: Mentor-Mentee Matching API
  description: API for matching mentors and mentees in a mentoring platform
  version: 1.0.0
  contact:
    name: Mentor-Mentee Matching App
  license:
    name: MIT
servers:
  - url: http://localhost:8088/api
    description: Local development server
security:
  - BearerAuth: []
paths:
  /signup:
    post:
      operationId: signup
      tags:
        - Authentication
      summary: User registration
      description: Register a new user as either a mentor or mentee
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User successfully created
        '400':
          description: Bad request - invalid payload format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      operationId: login
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request - invalid payload format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /feedback:
    post:
      operationId: createFeedback
      tags:
        - Feedback
      summary: Create feedback for a completed match
      description: Submit feedback and rating for a mentor or mentee after a successful match
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreate'
      responses:
        '201':
          description: Feedback created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          description: Bad request - invalid data or feedback already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not authorized to give feedback for this match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Match request not found or not accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /feedback/received:
    get:
      operationId: getReceivedFeedback
      tags:
        - Feedback
      summary: Get feedback received by current user
      description: Retrieve all feedback received by the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Feedback list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
        '401':
          description: Unauthorized - authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
        - name
        - role
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 6
          example: "password123"
        name:
          type: string
          example: "김멘토"
        role:
          type: string
          enum: [mentor, mentee]
          example: "mentor"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "password123"
    LoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    FeedbackCreate:
      type: object
      required:
        - matchRequestId
        - revieweeId
        - rating
      properties:
        matchRequestId:
          type: integer
          example: 1
          description: ID of the accepted match request
        revieweeId:
          type: integer
          example: 4
          description: ID of the user being reviewed
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
          description: Rating from 1 to 5 stars
        comment:
          type: string
          maxLength: 1000
          example: "Great mentor! Very helpful and knowledgeable."
          description: Optional feedback comment
    FeedbackResponse:
      type: object
      required:
        - id
        - matchRequestId
        - reviewerId
        - revieweeId
        - rating
      properties:
        id:
          type: integer
          example: 1
        matchRequestId:
          type: integer
          example: 1
        reviewerId:
          type: integer
          example: 2
        revieweeId:
          type: integer
          example: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Great mentor! Very helpful and knowledgeable."
    Feedback:
      type: object
      required:
        - id
        - matchRequestId
        - rating
        - createdAt
        - reviewer
      properties:
        id:
          type: integer
          example: 1
        matchRequestId:
          type: integer
          example: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Great mentor! Very helpful and knowledgeable."
        createdAt:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"
        reviewer:
          type: object
          required:
            - name
            - role
          properties:
            name:
              type: string
              example: "John Doe"
            role:
              type: string
              enum: [mentor, mentee]
              example: "mentor"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Error message"
        details:
          type: string
          example: "Detailed error information"
...